+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix bug in schema handling (compressor settings ignored)
 - Move making ICF field partition directories into per-partition processing.
   Remove progress on the init mkdirs step.
+- Turn off progress monitor on dexplode-partition
 
 # 0.0.4 2024-04-08