+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Move making ICF field partition directories into per-partition processing.
   Remove progress on the init mkdirs step.
 - Turn off progress monitor on dexplode-partition
+- Fix empty partition bug
 
 # 0.0.4 2024-04-08